# Speed up Rust tests by turning doctests into unit tests