--- a/rust/libnewsboat/src/utils.rs
+++ b/rust/libnewsboat/src/utils.rs
@@ -289,13 +289,6 @@ pub fn strwidth(rs_str: &str) -> usize {
 ///
 /// STFL tags (e.g. `<b>`, `<foobar>`, `</>`) are counted as having 0 width.
 /// Escaped less-than sign (`<` escaped as `<>`) is counted as having a width of 1 character.
-/// ```
-/// use libnewsboat::utils::strwidth_stfl;
-/// assert_eq!(strwidth_stfl("a"), 1);
-/// assert_eq!(strwidth_stfl("abc<tag>def"), 6);
-/// assert_eq!(strwidth_stfl("less-than: <>"), 12);
-/// assert_eq!(strwidth_stfl("ＡＢＣＤＥＦ"), 12);
-///```
 pub fn strwidth_stfl(mut s: &str) -> usize {
     let mut width = 0;
     loop {
@@ -1321,6 +1314,10 @@ mod tests {
         assert_eq!(strwidth_stfl("\u{F91F}"), 2);
         assert_eq!(strwidth_stfl("\u{0007}"), 0);
         assert_eq!(strwidth_stfl("<a"), 0); // #415
+        assert_eq!(strwidth_stfl("a"), 1);
+        assert_eq!(strwidth_stfl("abc<tag>def"), 6);
+        assert_eq!(strwidth_stfl("less-than: <>"), 12);
+        assert_eq!(strwidth_stfl("ＡＢＣＤＥＦ"), 12);
     }
 
     #[test]