--- a/rust/libnewsboat/src/utils.rs
+++ b/rust/libnewsboat/src/utils.rs
@@ -109,11 +109,6 @@ pub fn is_special_url(url: &str) -> bool {
 }
 
 /// Check if the given URL is a http(s) URL
-/// # Example
-/// ```
-/// use libnewsboat::utils::is_http_url;
-/// assert!(is_http_url("http://example.com"));
-/// ```
 pub fn is_http_url(url: &str) -> bool {
     url.starts_with("https://") || url.starts_with("http://")
 }
@@ -1200,6 +1195,7 @@ mod tests {
 
     #[test]
     fn t_is_http_url() {
+        assert!(is_http_url("http://example.com"));
         assert!(is_http_url("https://foo.bar"));
         assert!(is_http_url("http://"));
         assert!(is_http_url("https://"));