--- a/rust/libnewsboat/src/utils.rs
+++ b/rust/libnewsboat/src/utils.rs
@@ -551,18 +551,6 @@ pub fn make_title(rs_str: String) -> String {
 
 /// Run the given command interactively with inherited stdin and stdout/stderr. Return the lowest
 /// 8 bits of its exit code, or `None` if the command failed to start.
-/// ```
-/// use libnewsboat::utils::run_interactively;
-///
-/// let result = run_interactively("echo true", "test");
-/// assert_eq!(result, Some(0));
-///
-/// let result = run_interactively("exit 1", "test");
-/// assert_eq!(result, Some(1));
-///
-/// // Unfortunately, there is no easy way to provoke this function to return `None`, nor to test
-/// // that it returns just the lowest 8 bits.
-/// ```
 pub fn run_interactively(command: &str, caller: &str) -> Option<u8> {
     log!(Level::Debug, &format!("{}: running `{}'", caller, command));
     Command::new("sh")
@@ -2076,6 +2064,18 @@ mod tests {
             "http://test:test@foobar:33/bla2.html".to_owned()
         );
         assert_eq!(absolute_url("foo", "bar"), "bar".to_owned());
+    }
+
+    #[test]
+    fn t_run_interactively() {
+        let result = run_interactively("echo true", "test");
+        assert_eq!(result, Some(0));
+
+        let result = run_interactively("exit 1", "test");
+        assert_eq!(result, Some(1));
+
+        // Unfortunately, there is no easy way to provoke this function to return `None`, nor to test
+        // that it returns just the lowest 8 bits.
     }
 
     #[test]