For the rationale, please see e4f99a6ed4ca90dccf7c47aae1d436a5bdb9ad7e. If that doesn't convince you, leave a comment below :)

Some numbers:

```
$ git checkout --quiet d9580c8cf97d276e74ed6c1c805f368f83d469d2 \
  && cargo test --no-run \
  && sync \
  && TMPDIR=/dev/shm hyperfine --warmup 5 --min-runs 20 --command-name 'Prior to these changes' 'cargo test' \
  && git checkout --quiet e4f99a6ed4ca90dccf7c47aae1d436a5bdb9ad7e \
  && cargo test --no-run \
  && sync \
  && TMPDIR=/dev/shm hyperfine --warmup 5 --min-runs 20 --command-name 'After these changes' 'cargo test'
   Compiling libnewsboat v2.23.0 (/home/minoru/src/newsboat/rust/libnewsboat)
   Compiling libnewsboat-ffi v2.23.0 (/home/minoru/src/newsboat/rust/libnewsboat-ffi)
    Finished test [unoptimized + debuginfo] target(s) in 4.88s
Benchmark #1: Prior to these changes
  Time (mean ± σ):      5.331 s ±  0.022 s    [User: 26.051 s, System: 2.829 s]
  Range (min … max):    5.291 s …  5.380 s    20 runs

   Compiling libnewsboat v2.23.0 (/home/minoru/src/newsboat/rust/libnewsboat)
   Compiling libnewsboat-ffi v2.23.0 (/home/minoru/src/newsboat/rust/libnewsboat-ffi)
    Finished test [unoptimized + debuginfo] target(s) in 4.95s
Benchmark #1: After these changes
  Time (mean ± σ):      1.503 s ±  0.018 s    [User: 3.847 s, System: 0.239 s]
  Range (min … max):    1.473 s …  1.535 s    20 runs
```

That's with Intel Core i7-7700HQ, 4 cores, 8 threads. The working directory is on an SSD; `sync` ensures that all the executables are written onto disk before the testing starts. I decided against using tmpfs for the working directory because I wanted numbers that are close to real life. The first `git checkout` moves to `master` at the time of writing.

So yeah, (1-1.503/5.331)*100 = 72% reduction in runtime. Even if we take re-compilation into account, that's still (1 - (4.95+1.503)/(4.95+5.331))*100 = 37% reduction, which isn't great but is already noticeable.

Reviews are welcome! Don't fear the amount of commits — they're tiny, and purely to help with reviewing. I'll squash them when I merge this in three days.